+++ gwcelery/tasks/external_triggers.py
@@ -86,8 +86,15 @@ def handle_lvalert(alert):
     # Determine GraceDb ID
     graceid = alert['uid']
 
-    if alert['alert_type'] == 'new':
-        raven.coincidence_search(graceid, alert['object']).delay()
+    if alert['alert_type'] == 'new' and \
+            alert['object'].get('group', '') == 'External':
+        raven.coincidence_search(graceid, alert['object'], group='CBC').delay()
+        raven.coincidence_search(graceid, alert['object'],
+                                 group='Burst').delay()
+    elif alert['alert_type'] == 'new' and graceid.startswith('S'):
+        preferred_event_id = alert['object']['preferred_event']
+        group = gracedb.get_event(preferred_event_id)['group']
+        raven.coincidence_search(graceid, alert['object'], group=group).delay()
     elif alert['alert_type'] == 'label':
         if 'EM_COINC' in alert['description'] and graceid.startswith('S'):
             ligo_fermi_skymaps.create_combined_skymap(graceid).delay()